# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Viper Chain API
  version: "0.2.0"
  description: >
    API for interacting with Viper PQ Chain — a post-quantum digital trust
    L1. The public chain is `viper-testnet-1` (`chain_id_hex` is the hex of
    the ASCII chain id, e.g. `76697065722d746573746e65742d31`); it is
    created after the public release, so there is no live network to
    query at publication time. Routes marked "specified, not implemented"
    are documented for SDK alignment but are not served by `pqcd`. Use these
    endpoints to query chain state, submit transactions, look up
    attestations / proofs / governance receipts, manage verifiable
    credentials, and anchor proofs.
    Spec version aligns with the Viper SDK 0.2.0 release (TS + Python),
    which predates `viper-testnet-1` and is re-aligned before republishing.
servers:
  - url: https://rpc.pqchain.agwswebconsulting.it
    description: Planned public read API of viper-testnet-1 (DNS published at genesis; not live at publication time)
  - url: http://localhost:26657
    description: Local devnet node (default API port)
...
paths:
...
  /v1/blocks/latest:
    get:
      tags: [Chain]
      summary: Get the latest committed block
      description: >
        Returns the latest committed block header. Implemented by
        `crates/pqcd/src/api.rs::handle_latest_block`.
      operationId: getLatestBlock
      responses:
        "200":
          description: Latest block header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockResponse"
        "404":
          description: No blocks committed yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/txs/{hash}:
    get:
      tags: [Chain]
      summary: Get transaction by hash
      description: Looks up a finalized transaction by its hash (64-character hex string).
      operationId: getTransaction
      parameters:
        - name: hash
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-f]{64}$"
          description: Transaction hash (64 hex characters)
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          description: Invalid hash format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Transaction not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/accounts/{address}:
    get:
      tags: [Accounts]
      summary: Get account by address
      description: Returns account balance and nonce for the given address (64-character hex).
      operationId: getAccount
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-f]{64}$"
          description: Account address (64 hex characters)
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
              example:
                address: "087024f943f46283fbfffd2536313d74a87c39aee943f5b4dce88a6f1ba53cfc"
                nonce: 2
                balance: "49900"
        "400":
          description: Invalid address format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found
...
  /v1/attestations/{id}:
    get:
      tags: [Attestations]
      summary: Get attestation by id
      description: >
        Returns a finalized attestation record by its 32-byte id (64
        hex characters). Implemented by the node runtime at
        `crates/pqcd/src/devnet.rs::handle_attestation` and by the
        api-only handler at `crates/pqcd/src/api.rs::handle_attestation`.
      operationId: getAttestation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-f]{64}$"
          description: Attestation id (64 hex characters)
      responses:
        "200":
          description: Attestation details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttestationResponse"
        "400":
          description: Invalid attestation id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Attestation not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Proofs ───────────────────────────────────────────────────────────────────
  /v1/proofs/{anchor_id}:
    get:
      tags: [Proofs]
      summary: Get proof anchor by anchor id
      description: >
        Returns a finalized proof anchor record by its 32-byte anchor
        id (64 hex characters). Implemented by the node runtime at
        `crates/pqcd/src/devnet.rs::handle_get_proof_anchor` and by the
        api-only handler at `crates/pqcd/src/api.rs::handle_proof_anchor`.
      operationId: getProofAnchor
      parameters:
        - name: anchor_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-f]{64}$"
          description: Anchor id (64 hex characters)
      responses:
        "200":
          description: Proof anchor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProofAnchorResponse"
        "400":
          description: Invalid anchor id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Proof anchor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/proofs/{id}:
    get:
      tags: [Proofs]
      summary: Get proof by id
      description: >
        Returns a proof anchor record with business-friendly field names.
        Wraps the on-chain proof anchor data.
      operationId: getProof
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-f]{64}$"
          description: Proof ID (64 hex characters)
      responses:
        "200":
          description: Proof details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProofResponse"
        "400":
          description: Invalid proof id format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Proof not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Governance ───────────────────────────────────────────────────────────────
  /v1/governance/receipts/{proposal_id}:
    get:
      tags: [Governance]
      summary: Get governance proposal execution receipt (specified, not implemented)
      description: >
        Returns the execution receipt for a finalized governance
        proposal. Implemented by
        `crates/pqcd/src/api.rs::handle_governance_receipt`.
      operationId: getGovernanceReceipt
      parameters:
        - name: proposal_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-f]{64}$"
          description: Proposal id (tx hash, 64 hex characters)
      responses:
        "200":
          description: Governance receipt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GovernanceReceiptResponse"
        "400":
          description: Invalid proposal id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Governance receipt not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error description
          required: [code, message]

    # ── Chain ──────────────────────────────────────────────────────────────────
...
    BlockResponse:
      type: object
      properties:
        height:
          type: integer
          format: int64
        block_hash:
          type: string
        prev_hash:
          type: string
        state_root:
          type: string
        tx_root:
          type: string
        proposer:
          type: string
        timestamp:
          type: integer
          format: int64
          description: >
            Block timestamp. ADR-053 §T1.1 defines on-chain timestamps
            as uint64 nanoseconds UTC; the explorer surface continues
            to expose seconds for backwards compatibility, see
            CHANGELOG.md.
        tx_count:
          type: integer
        tx_hashes:
          type: array
          items:
            type: string
        bytes_used:
          type: integer
          format: int64
    TransactionResponse:
      type: object
      description: >
        Live shape from `crates/pqcd/src/devnet.rs::handle_tx_lookup`.
        Includes both the canonical fields (tx_hash, block_height,
        sender, op_type, nonce, fee_venom, fee_tip, alg_id) and the
        signature/payload bytes for replay verification.
      properties:
        tx_hash:
          type: string
        block_height:
          type: integer
          format: int64
        sender:
          type: string
        op_type:
          type: string
          description: Debug-formatted MsgType
        nonce:
          type: integer
        fee_venom:
          type: string
          description: Declared fee budget in venom (string to avoid u64 precision loss)
        fee_tip:
          type: integer
        alg_id:
          type: string
        signature:
          type: string
        op_payload:
          type: string
        status:
          type: string
          enum: [pending, finalized, failed, not_found]
...
    AccountResponse:
      type: object
      properties:
        address:
          type: string
        nonce:
          type: integer
        balance:
          type: string
          description: Balance as decimal string

    # ── Validators ─────────────────────────────────────────────────────────────
...
    AttestationResponse:
      type: object
      properties:
        attestation_id:
          type: string
        issuer:
          type: string
        subject:
          type: string
        schema_id:
          type: string
        payload_hash:
          type: string
        issued_at_height:
          type: integer
          format: int64
        revoked_at_height:
          type: integer
          format: int64
          nullable: true
        attestation_type:
          type: integer

    # ── Proofs ─────────────────────────────────────────────────────────────────
    ProofAnchorData:
      type: object
      properties:
        anchor_id:
          type: string
        claimer:
          type: string
        claim_type:
          type: integer
        claim_type_name:
          type: string
          nullable: true
        asset_id_hash:
          type: string
        proof_hash:
          type: string
        schema_id:
          type: string
          nullable: true
        anchor_height:
          type: integer
          format: int64
    ProofAnchorResponse:
      type: object
      description: Wrapper returned by GET /v1/proofs/{anchor_id}.
      properties:
        data:
          $ref: "#/components/schemas/ProofAnchorData"
...
    ProofResponse:
      type: object
      properties:
        proof_id:
          type: string
        owner:
          type: string
        claim_type:
          type: string
        document_hash:
          type: string
        proof_hash:
          type: string
        anchored_at_block:
          type: integer
          format: int64
        status:
          type: string
          enum: [anchored]
        verify_url:
          type: string

    # ── Governance ─────────────────────────────────────────────────────────────
    GovernanceReceiptResponse:
      type: object
      properties:
        proposal_id:
          type: string
        proposal_type:
          type: integer
        proposal_type_name:
          type: string
        proposer:
          type: string
        target_alg_id:
          type: integer
        lifecycle_before:
          type: string
          enum: [active, discouraged, deprecated, banned]
        lifecycle_after:
          type: string
          enum: [active, discouraged, deprecated, banned]
        min_fee_before:
          type: integer
          format: int64
        min_fee_after:
          type: integer
          format: int64
        rationale_hash:
          type: string
        executed_at_height:
          type: integer
          format: int64

    # ── Archival ───────────────────────────────────────────────────────────────
...